# docs/swagger.yaml
definitions:
...
  handlers.CreateRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role_name:
        type: string
      telephone:
        type: string
    type: object
  handlers.CuratorResponse:
    properties:
      course_ids:
        items:
          type: string
        type: array
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role_id:
        type: string
      student_ids:
        items:
          type: string
        type: array
      telephone:
        type: string
    type: object
...
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
  models.User:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      passwordHash:
        type: string
      resetTokenExpiresAt:
        type: string
      roleID:
        type: string
      telephone:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /settings/users:
    get:
      description: Возвращает список всех пользователей с возможностью фильтрации
        по роли
      parameters:
      - description: Фильтр по ID роли
        format: uuid
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Неверный формат UUID
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить список пользователей
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с указанной ролью. Для роли 'curator'
        автоматически создает связанную запись.
      parameters:
      - description: Данные для создания пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/models.ApiError'
        "409":
          description: Email уже существует
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Создать пользователя
      tags:
      - Users
  /settings/users/{id}:
    delete:
      description: Удаляет пользователя из системы
      parameters:
      - description: ID пользователя
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Пользователь удален
        "400":
          description: Неверный формат UUID
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Удалить пользователя
      tags:
      - Users
  /settings/users/{userId}:
    get:
      description: Возвращает информацию о пользователе по его UUID
      parameters:
      - description: ID пользователя
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Неверный формат UUID
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить пользователя по ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующем пользователе
      parameters:
      - description: ID пользователя
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Обновленные данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Данные обновлены
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Обновить пользователя
      tags:
      - Users
  /settings/users/curators:
    get:
      description: Возвращает список всех кураторов с дополнительной информацией (студенты
        и курсы)
      produces:
      - application/json
      responses:
        "200":
          description: Список кураторов с деталями
          schema:
            items:
              $ref: '#/definitions/handlers.CuratorResponse'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить список кураторов
      tags:
      - Users
  /settings/users/managers:
    get:
      description: Возвращает список всех пользователей с ролью 'manager'
      produces:
      - application/json
      responses:
        "200":
          description: Список менеджеров
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить список менеджеров
      tags:
      - Users
swagger: "2.0"